unify: use compositional data structure instead of inheritance
