+++ base/src/main/java/org/aya/tyck/error/UnifyError.java
@@ -16,65 +16,18 @@ import org.aya.util.error.SourcePos;
 import org.jetbrains.annotations.NotNull;
 
 public interface UnifyError extends TyckError {
-  @NotNull TyckState state();
-  @NotNull Unifier.FailureData failureData();
-
-  default @NotNull Doc describeUnify(
-    @NotNull PrettierOptions options,
-    @NotNull Doc prologue,
-    @NotNull Term actual,
-    @NotNull Doc epilogue,
-    @NotNull Term expected
-  ) {
-    var actualDoc = actual.toDoc(options);
-    var expectedDoc = expected.toDoc(options);
-    var actualNFDoc = nf(actual).toDoc(options);
-    var expectedNFDoc = nf(expected).toDoc(options);
-    var buf = MutableList.of(prologue);
-    compareExprs(epilogue, actualDoc, expectedDoc, actualNFDoc, expectedNFDoc, buf);
-    var failureTermL = failureData().lhs();
-    var failureTermR = failureData().rhs();
-    var failureLhs = failureTermL.toDoc(options);
-    if (!failureLhs.equals(actualDoc)
-      && !failureLhs.equals(expectedDoc)
-      && !failureLhs.equals(actualNFDoc)
-      && !failureLhs.equals(expectedNFDoc)
-    ) {
-      buf.append(Doc.english("In particular, we failed to unify"));
-      compareExprs(Doc.plain("with"),
-        failureLhs, failureTermR.toDoc(options),
-        nf(failureTermL).toDoc(options),
-        nf(failureTermR).toDoc(options),
-        buf);
-    }
-    return Doc.vcat(buf);
-  }
-  @NotNull private Term nf(Term failureTermL) {
-    return failureTermL.normalize(state(), NormalizeMode.NF);
-  }
-  private static void compareExprs(@NotNull Doc mid, Doc actualDoc, Doc expectedDoc, Doc actualNFDoc, Doc expectedNFDoc, MutableList<@NotNull Doc> buf) {
-    buf.append(Doc.par(1, actualDoc));
-    if (!actualNFDoc.equals(actualDoc))
-      buf.append(Doc.par(1, Doc.parened(Doc.sep(Doc.plain("Normalized:"), actualNFDoc))));
-    buf.append(mid);
-    buf.append(Doc.par(1, expectedDoc));
-    if (!expectedNFDoc.equals(expectedDoc))
-      buf.append(Doc.par(1, Doc.parened(Doc.sep(Doc.plain("Normalized:"), expectedNFDoc))));
-  }
-
   record Type(
     @Override @NotNull Expr expr,
     @NotNull Term expected,
     @NotNull Term actual,
-    @Override @NotNull Unifier.FailureData failureData,
-    @Override @NotNull TyckState state
+    @NotNull UnifyInfo info
   ) implements ExprProblem, UnifyError {
     @Override public @NotNull Doc describe(@NotNull PrettierOptions options) {
       var prologue = Doc.vcat(
         Doc.english("Cannot check the expression"),
         Doc.par(1, expr.toDoc(options)),
         Doc.english("of type"));
-      return describeUnify(options, prologue, actual, Doc.english("against the type"), expected);
+      return info.describeUnify(options, prologue, actual, Doc.english("against the type"), expected);
     }
   }
 
@@ -82,8 +35,7 @@ public interface UnifyError extends TyckError {
     @NotNull TeleDecl.DataCtor ctor,
     @NotNull Term expected,
     @NotNull Term actual,
-    @Override @NotNull Unifier.FailureData failureData,
-    @Override @NotNull TyckState state
+    @NotNull UnifyInfo info
   ) implements UnifyError {
     @Override public @NotNull SourcePos sourcePos() {
       return ctor.sourcePos;
@@ -94,7 +46,7 @@ public interface UnifyError extends TyckError {
         Doc.english("Cannot make sense of the return type of the constructor"),
         Doc.par(1, ctor.toDoc(options)),
         Doc.english("which eventually returns"));
-      return describeUnify(options, prologue, actual, Doc.english("while it should return"), expected);
+      return info.describeUnify(options, prologue, actual, Doc.english("while it should return"), expected);
     }
   }
 }
